streamline db and add permissions check

+++ src/handlers/v1/groups/delete-group.ts
@@ -1,6 +1,6 @@
 import { DynamoDBDocumentClient } from '@aws-sdk/lib-dynamodb';
 import { APIGatewayProxyEvent, APIGatewayProxyResult } from 'aws-lambda';
-import { deleteItem, getDbClient } from '../../../services/dynamodb';
+import { deleteItem, getDbClient, getItem, queryTable } from '../../../services/dynamodb';
 
 let dbClient: DynamoDBDocumentClient;
 
@@ -9,6 +9,7 @@ export const handler = async (event: APIGatewayProxyEvent): Promise<APIGatewayPr
 
   try {
     const groupId = event?.pathParameters?.groupId || false;
+    const user = event.requestContext.authorizer?.iam?.cognitoIdentity?.identityId || '';
 
     if (!groupId) {
       return {
@@ -26,14 +27,30 @@ export const handler = async (event: APIGatewayProxyEvent): Promise<APIGatewayPr
       dbClient = getDbClient();
     }
 
-    // TODO: Add permissions check
+    const groupSettings = (await getItem<GroupMetadata>(key, dbClient, ['settings'])).settings;
 
-    await deleteItem(key, dbClient);
+    if(groupSettings.administratorIds.includes(user)){
+      const attributes = ['pk', 'sk'];
+      const group = await queryTable<GroupMetadata|PlayLogEntry>(key.pk, 'ALL', dbClient, attributes);
 
-    response = {
-      statusCode: 200,
-      body: `Group ${groupId} deleted.`,
-    };
+      for (const item of group) {
+        const key = {
+          pk: item.pk,
+          sk: item.sk,
+        };
+        await deleteItem(key, dbClient);
+      }
+
+      response = {
+        statusCode: 200,
+        body: `Group ${groupId} deleted.`,
+      };
+    } else {
+      response = {
+        statusCode: 403,
+        body: 'Unauthorized.',
+      };
+    }
 
   } catch (err) {
     console.log(err);
